Fix resume and cover letter to open in a different tab
