+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@react-pdf/renderer": "^3.0.1",
     "eslint-config-next": "13.0.3",
     "framer-motion": "^7.6.7",
     "next": "13.0.3",
@@ -16,8 +17,7 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-icons": "^4.6.0",
-    "react-lottie": "^1.2.3",
-    "react-pdf": "^6.1.0"
+    "react-lottie": "^1.2.3"
   },
   "devDependencies": {
     "autoprefixer": "^10.4.13",
